NativeWindow public* reorg 3/3 ; NativeVisualID -> VisualIDHolder incl. proper utilization.
- VisualIDHolder: Update documentation (Exception case, etc)

- NativeVisualID -> VisualIDHolder (public)
  - incl. generic Comparator
  - better doc and enum values
  - VID_UNDEFINED == 0
  - methods shall not throw exception, but return UNDEFINED

- CapabilitiesImmutable extends VisualIDHolder

- All Capabilties impl.
  - use VID_UNDEFINED for undef. value
  - use final private (immutable) fields

- AbstractGraphicsConfiguration extends VisualIDHolder

- X11 CreateDummyWindow takes (int) visualID

+++ make/config/nativewindow/x11-CustomJavaCode.java
@@ -26,7 +26,7 @@
 
   public static native long DefaultVisualID(long display, int screen);
 
-  public static native long CreateDummyWindow(long display, int screen_index, long visualID, int width, int height);
+  public static native long CreateDummyWindow(long display, int screen_index, int visualID, int width, int height);
   public static native void DestroyDummyWindow(long display, long window);
 
   public static Point GetRelativeLocation(long display, int screen_index, long src_win, long dest_win, int src_x, int src_y) {
